do nothing if target file already exist

+++ setup.sh
@@ -8,5 +8,5 @@ cd ${BASE_DIR}
 for f in .??*
 do
   [[ "$f" == ".git" ]] && continue
-  ln -s "${BASE_DIR}/${f}" ${HOME}/
+  test -e "${HOME}/${f}" || ln -s "${BASE_DIR}/${f}" ${HOME}/
 done
